Use current state when fetching input artifacts
When fetching a process via the API, the input output map has all
artifacts listed with a particular state. We've historically used
this state flag when updating an artifact.

However, there was one exception to this. QC flags where being fetched
using no `state` flag when they were being set. The script developer had
to force this with an API call that was only available in
clarity-snpseq, i.e. the scripts used at SNP&SEQ, Uppsala University.

This patch fixes this so we now fetch without that state parameter.

Note that we had already fixed this in an earlier patch (d03a3d6 and
related) but it was done in a different way and this patch does some
extra refactorings and fixes as there was a bug in the earlier patch.